move all the env setup into `setup-env`
