+++ tests/e2e/jest.config.js
@@ -10,7 +10,6 @@ module.exports = {
 	setupFilesAfterEnv: [
 		'jest-allure/dist/setup',
 		'<rootDir>/lib/setup-env.js',
-		'<rootDir>/lib/jest.test.failure.js',
 		'expect-puppeteer',
 	],
 };
